pyproject.toml
[project]
name = "scroll"
version = "0.1.0"
description = ""
authors = [
    {name = "zhang qing",email = "zhangqing.opt@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4"
dependencies = [
    "django (>=5.2.1,<6.0.0)",
    "django-environ (>=0.12.0,<0.13.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)"
]

[tool.poetry]
packages = [{include = "scroll", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
